requirements.txt
fastapi
uvicorn
python-multipart
faiss-cpu
sentence-transformers
pypdf
langchain
numpy
streamlit
transformers
torch
